Migrating the database to XAMPP and fix a bug on New Student Form
After testing if the migration from MySQL to XAMPP is successful I noticed that there is a bug in the New Student Form wherein the program will close after trying to display the output from the database query.

The reason is that the query contains 19 items and the expected contents of the program is 18.

The Bug is now Fixed